requirements.txt
PyYAML~=6.0.1
APScheduler~=3.10.1
croniter~=2.0.1
python-telegram-bot~=20.8
python-telegram-bot[socks]
loguru~=0.7.2
httpx~=0.26.0
httpx[socks]
uvicorn~=0.27.0.post1
fastapi~=0.109.0
starlette~=0.35.1
prettytable
bencode-py
